 Removing all accounts changes the current account to a throwaway to signal returning to login activity.
First part of providing control over local message state added with an object holding the current "inboxes" the message is in. As it is, the state should be allowed to be null and it'll default to only being in the inbox. While I initially thought it would be better to hold these extra details in another table, I'm keeping it in the same table and letting updates from online override the local state. This is because after a message is read, there's only two ways a message's state will change: either it changes from read/unread to unread/read, or maybe if the correspondent deletes their account.